feat: handle records with multiple locations
When there is more than one location, the
'Access Online' button opens a dropdown allowing
the user to select with link to use based
on item location.

Ensure that you have more than one item for the
record used for testing, and that they are not
recorded as being found at the same location.

Test plan:

- search Aspen for the item you have added
- click on the search result's title or 'More
Info' button
- Scroll to the 'Links' section, and open it.
- Notice the doi link
- Notice the LibKey link
- Click the LibKey link
- if LibKey was able to
return a direct link to the text, this should take
you to a PDF of the document via LibKey, which
should open in a new tab.e 'Access Online' button
- On the search results page (/Union/Search),
open 'Show Edition(s)' on the record. Click the
'Access Online' link - select a location, click,
and notice that the redirect behaviour remains
the same.